# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: HaoHaoAccounting API
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    post:
      tags: [auth]
      operationId: postAuthLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '400':
          $ref: '#/components/responses/BadRequest'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /auth/refresh:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    post:
      tags: [auth]
      operationId: postAuthRefresh
      responses:
        '200':
          description: Refreshed token
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /auth/logout:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    post:
      tags: [auth]
      operationId: postAuthLogout
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /me:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    get:
      tags: [auth]
      operationId: getMe
      responses:
        '200':
          description: Current user
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /accounts:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    get:
      tags: [accounts]
      operationId: getAccounts
      responses:
        '200':
          description: Account list
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
    post:
      tags: [accounts]
      operationId: postAccounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: Created account
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /accounts/{id}:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    put:
      tags: [accounts]
      operationId: putAccountsById
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '200':
          description: Updated account
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
    delete:
      tags: [accounts]
      operationId: deleteAccountsById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Optional request correlation id. The API returns this value when valid, otherwise it generates one.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[!-~]+$'
  headers:
    RequestID:
      description: Request correlation id for tracing logs and API calls.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[!-~]+$'
...
    Vary:
      description: Request headers that influenced response representation selection.
      schema:
        type: string
    Allow:
      description: HTTP methods supported by the target route.
      schema:
        type: string
    WWWAuthenticate:
      description: Authentication challenge for protected resources. Invalid bearer tokens include error="invalid_token" per RFC 6750.
      schema:
        type: string
    RetryAfter:
      description: Retry delay as either HTTP-date or delay-seconds.
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request or validation error.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired authentication token.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Requested resource was not found.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowed:
      description: HTTP method is not supported for this route.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
        Allow:
          $ref: '#/components/headers/Allow'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Too many requests.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PayloadTooLarge:
      description: Request body is too large.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: Request Content-Type is not supported by this operation.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotAcceptable:
      description: Requested response media type is not available for this operation.
      headers:
        Vary:
          $ref: '#/components/headers/Vary'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GatewayTimeout:
      description: Request processing exceeded the server timeout budget.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Ok:
      description: OK response
      headers:
        Vary:
          $ref: '#/components/headers/Vary'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OkResponse'
  schemas:
...
    CurrentUser:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        phone:
          type: string
        email:
          type: string
        wechatId:
          type: string
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/CurrentUser'
    Account:
      type: object
      required: [id, userId, name, type, balance]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        type:
          type: string
        balance:
          type: number
    AccountRequest:
      type: object
      additionalProperties: false
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          minLength: 1
        balance:
          type: number
          minimum: 0
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [bad_request, invalid_request, unauthorized, forbidden, not_found, method_not_allowed, rate_limited, payload_too_large, unsupported_media_type, not_acceptable, request_timeout, client_closed_request, internal_error]
        requestId:
          type: string
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean